Skip controller tests for now, rename stubs
